Refactored the last packets.
tl;dr: The ProviderPacket, ResponsePacket, and ServiceCodePacket where updated with no changes the functionality.
ProviderPacket: Added a comment describing the job of this class and removed the extra namespaces not being used by this class.
ResponsePacket: Added a comment describing the job of this class and removed the extra namespaces not being used by this class. Also added a comment describing the ToString function.
ServiceCodePacket: Added a comment describing the job of this class and removed the extra namespaces not being used by this class.